Rename onewireentities.py to onewire_entities.py
