+++ homeassistant/components/onewire/onewire_entities.py
